Fixed a crash if ControlPanelSocket asks to getNodes more than once
This commit only comments out a line of code. That line of code for some reason runs before the line above it is done running. That ends up sending back the initial command back to ControlPanelSocket instead of sending back info depending on the initial command. The commented line does nothing, so removing it is enough.