Removed the 7ms temporary solution, implemented a real solution
The error was caused by a complex loop. The issue at hand is apparently that:
 - The server changes the state of all actuators
 - The server sends that information one line at a time to the client
 - The client gets the first line, and passes the information onwards through a listerner
 - The listener notifies the change, and therefore sends information to the server that there has been a change.

This should have ended up in an infinite loop, where the server continued to send actuator changes to the client, and the client continued to notify the listeners of a change, and therefore sending information to the server, notifying of a change. To then have the server inform the client of a change, and so on...
It is not a infinite loop however, since the "change" is not a change anymore after it is read by the client
ex:
- > the server turns on actuator 1
- > client gets notified of the change
- > the client notifies the server of the change, and tells the server that the actuator 1 is now on
- > the server sees that the change is in fact not a change, and doesn't do anything to continue the loop

The fix just makes sure that the client does not notify the change to the listeners *if* the change comes from the server, and not internally from the client itself